// apps/raise/src/commons/enums/allowed-filetype.enum.ts
export enum AllowedFileType {
  JPG = 'image/jpg',
  JPEG = 'image/jpeg',
  PNG = 'image/png',
  GIF = 'image/gif',
  SVG = 'image/svg+xml',
  MP4 = 'video/mp4',
  MKV = 'video/mkv',
  WMV = 'video/wmv',
  FLV = 'video/flv',
  AVI = 'video/avi',
  MPG = 'video/mpg',
  WEBM = 'video/webm',
  MOV = 'video/quicktime',
  MPEG = 'video/mpeg',
  MPEG2 = 'video/mpeg2',
  MPEG4 = 'video/mpeg4',
  OGGTHEORA = 'video/ogg',
  MP3 = 'audio/mpeg',
  OGG = 'audio/ogg',
  WAV = 'audio/wav',
  FLAC = 'audio/flac',
  OGGVORBIS = 'audio/ogg',
  OGGOPUS = 'audio/ogg',
  OGGSPEEX = 'audio/ogg',
  PDF = 'application/pdf',
  DOC = 'application/msword',
  DOCX = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
  XLS = 'application/vnd.ms-excel',
  XLSX = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
  CSV = 'text/csv',
  PPT = 'application/vnd.ms-powerpoint',
  PPTX = 'application/vnd.openxmlformats-officedocument.presentationml.presentation',
  ZIP = 'application/zip',
  XML = 'application/xml',
  JS = 'application/javascript',
  RAR = 'application/x-rar-compressed',
  JSON = 'application/json',
  TXT = 'text/plain',
  HTML = 'text/html',
  CSS = 'text/css',
}